Fix bugs in GKLeaderboard.LoadLeaderboards() and .LoadEntries() that caused them to either crash or return empty results. (#24)
* Fix bugs in GKLeaderboard.LoadLeaderboards() and .LoadEntries() that caused them to either crash or return empty results.
The main issue was that the downcasting helper NSObject.As<T>() assumed that the C# and Objective-C type names were the same. This was true in all cases except for GKLeaderboard.Entry (C#) and GKLeaderboardEntry (Obj-C). To avoid a breaking change that would be introduced by renaming the C# type to match, this change instead adds some special case code that allows specific types to provide a name, via reflection, that is used when matching against the corresponding Obj-C type.