rdar://143816043 (Dependency Confusion Vulnerability in Private NPM Package (com.unity.ai.navigation) Leading to RCE) (#63)
* rdar://143816043 (Dependency Confusion Vulnerability in Private NPM Package (com.unity.ai.navigation) Leading to RCE)
Remove many references to unused packages. This is in response to a reported supply-chain attack on com.unity.ai.navigation on npmjs.com. That repository isn't used by the Unity Editor to obtain first-party Unity packages, but this seemed like a good excuse to pare down our references to unused packages.

* Increment plug-in version numbers and update changelogs.