rdar://161912540 ([GameKitUnity][API Question for Testing] GKChallengeDefinition -> durationOptions) (#77)
The test app was using a filter for `ValidDate` and then displaying the
Date. This caused two problems

1. The dates provided in the challenges weren't Valid. In general this
   would be wrong because the Calendar property must be set for this
   ever to be true. See https://developer.apple.com/documentation/foundation/datecomponents/isvaliddate
2. The NSDateComponents had a day and not a Date.

There's probably more refinement that could be done here but this fixes
the immediate issue. We might need to provide better interop with
NSDateComponents in the future if they are intended to be used in a
similar fashion as DateTime in C#